mkfs.zfs: Added swap, fixed a typo.
If you set up zfs root, and try to create a swapfile, it fails because
ZFS detects you writing a file of all zeroes and automatically turns it
into a sparse file, which causes swapon to fail.

The ZFS way to do that is to create a volume and use that as swap, which
actually works in Linux.  So here we go.  I arbitrarily set the swap
volume size to be twice the size of RAM.

Also, if you're setting up ZFS as not root, I fixed the path of the zfs
command which was wrong.