Stage 1 modules: Ensure that gawk gets built before bash
This avoids having awk opportunistically link with libreadline,
which it does so automatically if it's available.  This kills the
ISO build because, when libncurses gets upgraded, the previous
version is removed, which stops awk from working.  So build awk
earlier enough that it automatically doesn't link with
libreadline, so that it'll continue working while libncurses is
unavailable.